+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Change Log
 ----------
 
+
+Development
+~~~~~~~~~~~
+
+* make subitems in the sidebar menu collapsible
+
+
 1.0.2
 ~~~~~
 
@@ -11,6 +18,7 @@ Change Log
 * ability to create new devices from the web application
 
 * several minor bugfixes
+
 
 1.0.1
 ~~~~~
@@ -25,6 +33,7 @@ Change Log
 * minor fixes for the SQLite backend
 
 * minor documentation fixes and updates
+
 
 1.0.0
 ~~~~~